some fixes to purge

+++ commands/purge.js
@@ -2,9 +2,16 @@ exports.run = (bot, message, args, level) => {
     var num = parseInt(args[0]);
     num += 1;
 	if (!isNaN(num)) {
-        message.channel.bulkDelete(num);
-        message.channel.send(`${args[0]} messages purged by ${message.author}!`)
-			.then(msg => setTimeout(function() {msg.delete()}, 5000));
+        if (num < 2) {
+            message.reply('Please input a number above 0! Thank you!')
+        } else {
+            if (num > 100) {
+                num = 100;
+            }
+            message.channel.bulkDelete(num);
+            message.channel.send(`${args[0]} messages purged by ${message.author}!`)
+			    .then(msg => setTimeout(function() {msg.delete()}, 5000));
+        }
 	} else {
 		message.channel.sendMessage("Please specify a number!");
 	}
@@ -14,7 +21,7 @@ exports.conf = {
     enabled: true,
     guildOnly: false,
     aliases: ['clear', 'prune', 'clean'],
-    permLevel: 2
+    permLevel: 4
 };
 
 exports.help = {
